+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Scatter Plot: Improve support for user-defined line widths (#2739, #2750) _Thanks @dayo05_
 * Fonts: Users can now assign `Fonts.Default` to set the font for all labels (#2753)
 * Fonts: Added `Fonts.Detect()` to facilitate selecting fonts that support international characters (#2746) _Thanks @heartacker_
+* Axis: Exposed `TickFont` to allow tick label size and style customization (#2747) _Thanks @heartacker_
+* Plot: Added `Title()`, `XLabel()`, and `YLabel()` helper methods
 
 ## ScottPlot 4.1.66 (in development)
 * DataLogger: Improved support for single-point datasets (#2733) _Thanks @KroMignon_